api_flask/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["mat <mateusz.baranczyk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.1"
cryptography = "^43.0.0"
gunicorn = "^23.0.0"

[tool.poetry.dev-dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
